test(editor): slim movement routing coverage
